adds state to form view
