+++ src/view/new-point-button-view.js
@@ -1,8 +1,6 @@
 import AbstractView from '../framework/view/abstract-view';
 
-function createNewPointButtonTemplate() {
-  return '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
-}
+const createNewPointButtonTemplate = () => '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
 
 export default class NewPointButtonView extends AbstractView {
   get template() {
