+++ src/view/trip-points-list-view.js
@@ -1,8 +1,6 @@
 import AbstractView from '../framework/view/abstract-view';
 
-function createTripPointsListTemplate() {
-  return '<ul class="trip-events__list"></ul>';
-}
+const createTripPointsListTemplate = () => '<ul class="trip-events__list"></ul>';
 
 export default class TripPointsListView extends AbstractView {
   get template() {
